Add files via upload
Transfer learning is a part of machine learning where a model proposed for one task can be used to solve the problem of another task. Deep learning models learn the weights and features that are necessary for the classiﬁcation of images. These deep learning models can be reused for the problem of a similar type. The layers at the beginning of the network which are used for feature extraction are reused but the lower layers which are too task-speciﬁc are changed. One of the main uses of transfer learning is in the ﬁeld of computer vision. Various algorithms had been proposed in computer vision that can be used for classiﬁcation of image on which network is not trained by modifying the lower layers. In this project, pre-trained Resnet50 is used for task of crack detection.